move ArrowUtils
