+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/pythonworker/ArrowUtils.scala
@@ -1,37 +1,33 @@
-package edu.uci.ics.texera.workflow.operators.udf.pythonV1
+package edu.uci.ics.amber.engine.architecture.pythonworker
 
 import edu.uci.ics.texera.workflow.common.tuple.Tuple
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils.AttributeTypeException
-import edu.uci.ics.texera.workflow.common.tuple.schema.{
-  Attribute,
-  AttributeType,
-  AttributeTypeUtils,
-  Schema
-}
-import org.apache.arrow.vector._
+import edu.uci.ics.texera.workflow.common.tuple.schema.{Attribute, AttributeType, AttributeTypeUtils, Schema}
 import org.apache.arrow.vector.types.FloatingPointPrecision
 import org.apache.arrow.vector.types.TimeUnit.MILLISECOND
 import org.apache.arrow.vector.types.pojo.ArrowType.PrimitiveType
 import org.apache.arrow.vector.types.pojo.{ArrowType, Field}
+import org.apache.arrow.vector.{BigIntVector, BitVector, FieldVector, Float8Vector, IntVector, TimeStampVector, VarCharVector, VectorSchemaRoot}
 
 import java.nio.charset.StandardCharsets
 import java.util
-import scala.collection.JavaConverters._
-import scala.collection.convert.ImplicitConversions.`list asScalaBuffer`
+import scala.collection.convert.ImplicitConversions.`collection AsScalaIterable`
+import scala.jdk.CollectionConverters.asJavaIterableConverter
 import scala.language.implicitConversions
+
 object ArrowUtils {
 
   implicit def bool2int(b: Boolean): Int = if (b) 1 else 0
 
   /**
     * Reads a row of the given Arrow Vectors into a Texera.Tuple
-    *    e.g.,
-    *    rowIndex  IntVector BigIntVector  BooleanVector
-    *    0         1         100L          true
+    * e.g.,
+    * rowIndex  IntVector BigIntVector  BooleanVector
+    * 0         1         100L          true
     *
-    *    the row at rowIndex 0 can be converted into `Tuple[1, 100L, true]`
+    * the row at rowIndex 0 can be converted into `Tuple[1, 100L, true]`
     *
-    * @param rowIndex The row index of the target row to be converted in the Vectors.
+    * @param rowIndex         The row index of the target row to be converted in the Vectors.
     * @param vectorSchemaRoot The root of the Vectors that stores the Arrow Fields. It contains multiple Vectors.
     * @return
     */
