+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/udf/pythonV1/PythonUDFOpExec.java
@@ -2,6 +2,7 @@ package edu.uci.ics.texera.workflow.operators.udf.pythonV1;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
+import edu.uci.ics.amber.engine.architecture.pythonworker.ArrowUtils;
 import edu.uci.ics.amber.engine.common.InputExhausted;
 import edu.uci.ics.amber.engine.common.virtualidentity.LinkIdentity;
 import edu.uci.ics.texera.Utils;
